backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
python-multipart
passlib[bcrypt]
python-dotenv
pymongo